requirements.txt
atlassian-python-api==3.41.4
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
Deprecated==1.2.14
html2text==2020.1.16
idna==3.6
oauthlib==3.2.2
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
urllib3==2.1.0
wrapt==1.16.0
atlassian-python-api==3.41.4
attrs==23.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
Deprecated==1.2.14
h11==0.14.0
html2text==2020.1.16
idna==3.6
oauthlib==3.2.2
outcome==1.3.0.post0
PySocks==1.7.1
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.2
trio-websocket==0.11.1
urllib3==2.1.0
wrapt==1.16.0
wsproto==1.2.0
